+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.2-alpha] - 2020-01-12
 ## Added
 - Added aliases for most commands (eg. `deploy,d`, `restart,rr`, `run,r`, `stop,s`).
 - Added two new options for logs `--all` and `--follow`.